simplify some packages
