+++ src/main/java/uk/gov/legislation/endpoints/contents/ContentsApi.java
@@ -1,4 +1,4 @@
-package uk.gov.legislation.endpoints.document.api;
+package uk.gov.legislation.endpoints.contents;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -11,23 +11,19 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import uk.gov.legislation.api.parameters.*;
 import uk.gov.legislation.api.parameters.Number;
-import uk.gov.legislation.api.responses.Document;
+import uk.gov.legislation.api.responses.TableOfContents;
 
 import java.util.Locale;
 import java.util.Optional;
 
-/**
- * API for document retrieval and transformation operations.
- */
-@Tag(name = "Documents")
-public interface DocumentApi {
+@Tag(name = "Tables of contents")
+public interface ContentsApi {
 
-    /**
-     * Retrieves document content in CLML format.
-     */
-    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "application/xml")
+    /* CLML */
+
+    @GetMapping(value = "/contents/{type}/{year}/{number}", produces = "application/xml")
     @Operation(
-        summary = "get a document with a calendar year",
+        summary = "get a document's table of conents (calendar year)",
         parameters = {
             @Parameter(
                 name = "Accept-Language",
@@ -37,16 +33,16 @@ public interface DocumentApi {
             )
         }
     )
-    ResponseEntity<String> getDocumentClml(
+    ResponseEntity<String> getDocumentContentsClml(
             @PathVariable @Type String type,
             @PathVariable @Year int year,
             @PathVariable @Number int number,
             @RequestParam @Version Optional<String> version,
             Locale locale) throws Exception;
 
-    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "application/xml")
+    @GetMapping(value = "/contents/{type}/{monarch}/{years}/{number}", produces = "application/xml")
     @Operation(
-        summary = "get a document with a regnal year",
+        summary = "get a document's table of conents (regnal year)",
         parameters = {
             @Parameter(
                 name = "Accept-Language",
@@ -56,7 +52,7 @@ public interface DocumentApi {
             )
         }
     )
-    ResponseEntity<String> getDocumentClml(
+    ResponseEntity<String> getDocumentContentsClml(
             @PathVariable @Type String type,
             @PathVariable @Monarch String monarch,
             @PathVariable @Years String years,
@@ -64,19 +60,18 @@ public interface DocumentApi {
             @RequestParam @Version Optional<String> version,
             Locale locale) throws Exception;
 
-    /**
-     * Retrieves document content in AKN format.
-     */
-    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "application/akn+xml")
-    ResponseEntity<String> getDocumentAkn(
+    /* Akoma Ntoso */
+
+    @GetMapping(value = "/contents/{type}/{year}/{number}", produces = "application/akn+xml")
+    ResponseEntity<String> getDocumentContentsAkn(
             @PathVariable String type,
             @PathVariable int year,
             @PathVariable int number,
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
 
-    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "application/akn+xml")
-    ResponseEntity<String> getDocumentAkn(
+    @GetMapping(value = "/contents/{type}/{monarch}/{years}/{number}", produces = "application/akn+xml")
+    ResponseEntity<String> getDocumentContentsAkn(
             @PathVariable String type,
             @PathVariable String monarch,
             @PathVariable String years,
@@ -84,50 +79,28 @@ public interface DocumentApi {
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
 
-    /**
-     * Retrieves document content in HTML format.
-     */
-    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "text/html")
-    ResponseEntity<String> getDocumentHtml(
+    /* JSON */
+
+    @GetMapping(value = "/contents/{type}/{year}/{number}", produces = "application/json")
+    ResponseEntity<TableOfContents> getDocumentContentsJson(
             @PathVariable String type,
             @PathVariable int year,
             @PathVariable int number,
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
 
-    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "text/html")
-    ResponseEntity<String> getDocumentHtml(
+    @GetMapping(value = "/contents/{type}/{monarch}/{years}/{number}", produces = "application/json")
+    ResponseEntity<TableOfContents> getDocumentContentsJson(
             @PathVariable String type,
             @PathVariable String monarch,
             @PathVariable String years,
             @PathVariable int number,
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
-
-    /**
-     * Retrieves document metadata and HTML content in JSON format.
-     */
-    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "application/json")
-    ResponseEntity<Document> getDocumentJson(
-            @PathVariable String type,
-            @PathVariable int year,
-            @PathVariable int number,
-            @RequestParam Optional<String> version,
-            Locale locale) throws Exception;
-
-    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "application/json")
-    ResponseEntity<Document> getDocumentJson(
-            @PathVariable String type,
-            @PathVariable String monarch,
-            @PathVariable String years,
-            @PathVariable int number,
-            @RequestParam Optional<String> version,
-            Locale locale) throws Exception;
-
 
     /* Word (.docx) */
 
-    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
+    @GetMapping(value = "/contents/{type}/{year}/{number}", produces = "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
     ResponseEntity<byte[]> docx(
             @PathVariable String type,
             @PathVariable int year,
@@ -135,7 +108,7 @@ public interface DocumentApi {
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
 
-    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
+    @GetMapping(value = "/contents/{type}/{monarch}/{years}/{number}", produces = "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
     ResponseEntity<byte[]> docx(
             @PathVariable String type,
             @PathVariable String monarch,
