+++ src/main/java/uk/gov/legislation/endpoints/document/DocumentApi.java
@@ -1,4 +1,4 @@
-package uk.gov.legislation.endpoints.contents.api;
+package uk.gov.legislation.endpoints.document;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -11,19 +11,23 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import uk.gov.legislation.api.parameters.*;
 import uk.gov.legislation.api.parameters.Number;
-import uk.gov.legislation.api.responses.TableOfContents;
+import uk.gov.legislation.api.responses.Document;
 
 import java.util.Locale;
 import java.util.Optional;
 
-@Tag(name = "Tables of contents")
-public interface ContentsApi {
+/**
+ * API for document retrieval and transformation operations.
+ */
+@Tag(name = "Documents")
+public interface DocumentApi {
 
-    /* CLML */
-
-    @GetMapping(value = "/contents/{type}/{year}/{number}", produces = "application/xml")
+    /**
+     * Retrieves document content in CLML format.
+     */
+    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "application/xml")
     @Operation(
-        summary = "get a document's table of conents (calendar year)",
+        summary = "get a document with a calendar year",
         parameters = {
             @Parameter(
                 name = "Accept-Language",
@@ -33,16 +37,16 @@ public interface ContentsApi {
             )
         }
     )
-    ResponseEntity<String> getDocumentContentsClml(
+    ResponseEntity<String> getDocumentClml(
             @PathVariable @Type String type,
             @PathVariable @Year int year,
             @PathVariable @Number int number,
             @RequestParam @Version Optional<String> version,
             Locale locale) throws Exception;
 
-    @GetMapping(value = "/contents/{type}/{monarch}/{years}/{number}", produces = "application/xml")
+    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "application/xml")
     @Operation(
-        summary = "get a document's table of conents (regnal year)",
+        summary = "get a document with a regnal year",
         parameters = {
             @Parameter(
                 name = "Accept-Language",
@@ -52,7 +56,7 @@ public interface ContentsApi {
             )
         }
     )
-    ResponseEntity<String> getDocumentContentsClml(
+    ResponseEntity<String> getDocumentClml(
             @PathVariable @Type String type,
             @PathVariable @Monarch String monarch,
             @PathVariable @Years String years,
@@ -60,18 +64,19 @@ public interface ContentsApi {
             @RequestParam @Version Optional<String> version,
             Locale locale) throws Exception;
 
-    /* Akoma Ntoso */
-
-    @GetMapping(value = "/contents/{type}/{year}/{number}", produces = "application/akn+xml")
-    ResponseEntity<String> getDocumentContentsAkn(
+    /**
+     * Retrieves document content in AKN format.
+     */
+    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "application/akn+xml")
+    ResponseEntity<String> getDocumentAkn(
             @PathVariable String type,
             @PathVariable int year,
             @PathVariable int number,
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
 
-    @GetMapping(value = "/contents/{type}/{monarch}/{years}/{number}", produces = "application/akn+xml")
-    ResponseEntity<String> getDocumentContentsAkn(
+    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "application/akn+xml")
+    ResponseEntity<String> getDocumentAkn(
             @PathVariable String type,
             @PathVariable String monarch,
             @PathVariable String years,
@@ -79,28 +84,50 @@ public interface ContentsApi {
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
 
-    /* JSON */
-
-    @GetMapping(value = "/contents/{type}/{year}/{number}", produces = "application/json")
-    ResponseEntity<TableOfContents> getDocumentContentsJson(
+    /**
+     * Retrieves document content in HTML format.
+     */
+    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "text/html")
+    ResponseEntity<String> getDocumentHtml(
             @PathVariable String type,
             @PathVariable int year,
             @PathVariable int number,
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
 
-    @GetMapping(value = "/contents/{type}/{monarch}/{years}/{number}", produces = "application/json")
-    ResponseEntity<TableOfContents> getDocumentContentsJson(
+    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "text/html")
+    ResponseEntity<String> getDocumentHtml(
             @PathVariable String type,
             @PathVariable String monarch,
             @PathVariable String years,
             @PathVariable int number,
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
+
+    /**
+     * Retrieves document metadata and HTML content in JSON format.
+     */
+    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "application/json")
+    ResponseEntity<Document> getDocumentJson(
+            @PathVariable String type,
+            @PathVariable int year,
+            @PathVariable int number,
+            @RequestParam Optional<String> version,
+            Locale locale) throws Exception;
+
+    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "application/json")
+    ResponseEntity<Document> getDocumentJson(
+            @PathVariable String type,
+            @PathVariable String monarch,
+            @PathVariable String years,
+            @PathVariable int number,
+            @RequestParam Optional<String> version,
+            Locale locale) throws Exception;
+
 
     /* Word (.docx) */
 
-    @GetMapping(value = "/contents/{type}/{year}/{number}", produces = "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
+    @GetMapping(value = "/document/{type}/{year}/{number}", produces = "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
     ResponseEntity<byte[]> docx(
             @PathVariable String type,
             @PathVariable int year,
@@ -108,7 +135,7 @@ public interface ContentsApi {
             @RequestParam Optional<String> version,
             Locale locale) throws Exception;
 
-    @GetMapping(value = "/contents/{type}/{monarch}/{years}/{number}", produces = "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
+    @GetMapping(value = "/document/{type}/{monarch}/{years}/{number}", produces = "application/vnd.openxmlformats-officedocument.wordprocessingml.document")
     ResponseEntity<byte[]> docx(
             @PathVariable String type,
             @PathVariable String monarch,
