+++ src/main/java/uk/gov/legislation/endpoints/fragment/FragmentController.java
@@ -1,4 +1,4 @@
-package uk.gov.legislation.endpoints.fragment.controller;
+package uk.gov.legislation.endpoints.fragment;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
@@ -6,20 +6,19 @@ import org.springframework.web.bind.annotation.RestController;
 import uk.gov.legislation.api.responses.Fragment;
 import uk.gov.legislation.data.marklogic.legislation.Legislation;
 import uk.gov.legislation.endpoints.CustomHeaders;
-import uk.gov.legislation.endpoints.document.controller.DocumentApiController.Transform;
-import uk.gov.legislation.endpoints.fragment.api.FragmentApi;
+import uk.gov.legislation.endpoints.document.DocumentController.Transform;
 import uk.gov.legislation.transform.Transforms;
 
 import java.util.Locale;
 import java.util.Optional;
 
 @RestController
-public class FragmentApiController implements FragmentApi {
+public class FragmentController implements FragmentApi {
 
     private final Legislation marklogic;
     private final Transforms transforms;
 
-    public FragmentApiController(Legislation marklogic, Transforms transforms) {
+    public FragmentController(Legislation marklogic, Transforms transforms) {
         this.marklogic = marklogic;
         this.transforms = transforms;
     }
